+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### 1.4.5
 - Updated Forge to 2296 (raoulvdberge)
+- Added Portable Grid (raoulvdberge)
+- Added OpenComputers integration (thraaawn)
 - Fixed Crafting Tweaks buttons positioned wrongly (blay09)
 - Fixed Crafting Tweaks keybindings interfering with RS keybindings (blay09)
 - Fixed crash when updating storages (raoulvdberge)